+++ minijinja/tests/test_value.rs
@@ -1,6 +1,7 @@
 use std::cmp::Ordering;
 
 use minijinja::value::Value;
+use minijinja::ErrorKind;
 
 #[test]
 fn test_sort() {
@@ -58,4 +59,26 @@ fn test_value_serialization() {
 fn test_float_to_string() {
     assert_eq!(Value::from(42.4242f64).to_string(), "42.4242");
     assert_eq!(Value::from(42.0f32).to_string(), "42.0");
+}
+
+#[test]
+fn test_value_as_slice() {
+    let val = Value::from(vec![1u32, 2, 3]);
+    assert_eq!(
+        val.as_slice().unwrap(),
+        &[Value::from(1), Value::from(2), Value::from(3)]
+    );
+    assert_eq!(Value::UNDEFINED.as_slice().unwrap(), &[]);
+    assert_eq!(Value::from(()).as_slice().unwrap(), &[]);
+    assert_eq!(
+        Value::from("foo").as_slice().unwrap_err().kind(),
+        ErrorKind::ImpossibleOperation
+    );
+}
+
+#[test]
+fn test_value_by_index() {
+    let val = Value::from(vec![1u32, 2, 3]);
+    assert_eq!(val.get_item_by_index(0).unwrap(), Value::from(1));
+    assert!(val.get_item_by_index(4).unwrap().is_undefined());
 }